requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cfgv==3.1.0
chardet==3.0.4
discord.py==1.3.4
distlib==0.3.1
filelock==3.0.12
identify==1.4.24
idna==2.10
multidict==4.7.6
nodeenv==1.4.0
pre-commit==2.6.0
python-dotenv==0.14.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
urllib3==1.25.9
virtualenv==20.0.27
websockets==8.1
yarl==1.4.2
